Add files via upload
This shell script will start the mystic internet service (mis). It checks to see if a stale semaphore is left behind and will also error out if the process is already running and a start attempt is made. This script can ideally be called as a fork within a systemd .service file. The script will exit with an appropriate error code to indicate whether is was successful or not. It is intended to work only with Debian, Ubuntu and other debian based distributions. For more information visit: https://vswitchzero.com/mystic-systemd