Add files via upload
This shell script will stop the mystic internet service (mis). It will fail out if the process is not running and a stop attempt is made. It will also continually poll to make sure it stops successfully, and if it is still running after 60 seconds, will be forcefully terminated. This script can ideally be called as a fork within a systemd .service file. The script will exit with an appropriate error code to indicate whether is was successful or not. It is intended to work only with Debian, Ubuntu and other debian based distributions. For more information visit: https://vswitchzero.com/mystic-systemd